Changes to WolfX
Can now interface with Yami-Yugi without needing to use shell.

Removed Card ID generator since I cbf doing regex to make the card names safe for a C enum, and currently there is no reason to have it. This will be reassessed if needed in future.

Mentally preparing for the coming things that need to be done.